+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/discovery/PolarisDiscoveryAutoConfiguration.java
@@ -18,13 +18,16 @@
 
 package com.tencent.cloud.polaris.discovery;
 
+import com.tencent.cloud.polaris.PolarisDiscoveryProperties;
 import com.tencent.cloud.polaris.discovery.reactive.PolarisReactiveDiscoveryClientConfiguration;
 import com.tencent.cloud.polaris.discovery.refresh.PolarisRefreshConfiguration;
+import com.tencent.cloud.polaris.extend.nacos.NacosContextProperties;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
+import org.springframework.lang.Nullable;
 
 /**
  * Discovery Auto Configuration for Polaris.
@@ -40,7 +43,12 @@ public class PolarisDiscoveryAutoConfiguration {
 	@Bean
 	@ConditionalOnMissingBean
 	public PolarisServiceDiscovery polarisServiceDiscovery(
+			@Nullable NacosContextProperties nacosContextProperties,
+			PolarisDiscoveryProperties polarisDiscoveryProperties,
 			PolarisDiscoveryHandler polarisDiscoveryHandler) {
-		return new PolarisServiceDiscovery(polarisDiscoveryHandler);
+		return new PolarisServiceDiscovery(
+				nacosContextProperties,
+				polarisDiscoveryProperties,
+				polarisDiscoveryHandler);
 	}
 }